[Grants] Refactor grants page into multiple components
